requirements.txt
streamlit>=1.28.0
pandas
python-dotenv
google-generativeai
langchain
langchain-google-genai
pytesseract
pdf2image
Pillow
